vector - compile with flags (but -std=c++98)

+++ srcs/vector/capacity.tpp
@@ -23,17 +23,17 @@ void	ft::vector<T, Alloc>::reserve(size_type new_cap)
 	if (new_cap <= this->_capacity) {return;}
 
 	T	*data = new T[this->_size];
-	for (int i = 0; i < this->_size; i++)
+	for (int i = 0; i < (int)this->_size; i++)
 		data[i] = this->_datas[i];
 
-	for (int i = 0; i < this->_size; i++)
+	for (int i = 0; i < (int)this->_size; i++)
 		this->_alloc.destroy(&this->_datas[i]);
 	this->_alloc.deallocate(this->_datas, this->_size);
 	
 	this->_capacity = new_cap;
 
 	this->_datas = this->_alloc.allocate(this->_capacity);
-	for (int i = 0; i < this->_size; i++)
+	for (int i = 0; i < (int)this->_size; i++)
 		this->_alloc.construct(&this->_datas[i], data[i]);
 	
 	delete[] data;
